chore: Refine Microservice Servers (#111)
* add `@NonNull` to `message` and `channel` `preSend()` args

* optimize build process for microservices

* fix grafana local ingress host

* add port 443 to grafana ingress local

* change `hourlyRate` from float to integer

* remove `rating` and `ratingSummary` for tutors

* run make api-generate

* remove float for minRate and maxRate

* cleanup stale generated python file

* adjust marketplace server test values

* make cert-manager null for grafana local values

+++ artifacts/server/marketplace/src/main/java/com/worksonourmachines/marketplace/tutorprofile/persistence/entity/MarketplaceTutorProfileEntity.java
@@ -38,7 +38,7 @@ public class MarketplaceTutorProfileEntity {
     private String bio;
 
     @Column(name = "hourly_rate", nullable = false)
-    private Float hourlyRate;
+    private Integer hourlyRate;
 
     @Column(name = "published", nullable = false)
     private Boolean published;
@@ -80,7 +80,7 @@ public class MarketplaceTutorProfileEntity {
             UUID userId,
             String displayName,
             String bio,
-            Float hourlyRate,
+            Integer hourlyRate,
             Boolean published) {
         this.userId = userId;
         this.displayName = displayName;
@@ -120,11 +120,11 @@ public class MarketplaceTutorProfileEntity {
         this.bio = bio;
     }
 
-    public Float getHourlyRate() {
+    public Integer getHourlyRate() {
         return hourlyRate;
     }
 
-    public void setHourlyRate(Float hourlyRate) {
+    public void setHourlyRate(Integer hourlyRate) {
         this.hourlyRate = hourlyRate;
     }
